LOG-151: added new node
