+++ include/rs485_port_manager/KillMissionRS485.hpp
@@ -18,7 +18,7 @@
 #include "sonia_common_ros2/msg/mission_status.hpp"
 #include "InterfaceModuleRS485.hpp"
 #include "sonia_common_ros2/msg/mission_status.hpp"
-#include "RS485Provider.hpp"
+#include "RS485Driver.hpp"
 
 namespace rs485_port_manager
 {
@@ -28,7 +28,7 @@ namespace rs485_port_manager
         
         public:
             KillMissionRS485();
-            ~KillMissionRS485();
+            ~KillMissionRS485() = default;
 
         /**
          * @brief method to send message to rs485
@@ -45,7 +45,7 @@ namespace rs485_port_manager
         void messageRS485CallBack(queueObject queue) override;
 
         private:
-        RS485Provider *rs485;
+        RS485Driver *rs485;
         /**
          * @brief Polls the status of kill and mission via RS485 connection.
          *
