+++ include/rs485_port_manager/PowerRS485.hpp
@@ -12,7 +12,7 @@
 #include "sonia_common_ros2/msg/motor_pwm.hpp"
 
 #include "InterfaceModuleRS485.hpp"
-#include "RS485Provider.hpp"
+#include "RS485Driver.hpp"
 
 namespace rs485_port_manager{
 
@@ -21,10 +21,10 @@ namespace rs485_port_manager{
         public:
 
             PowerRS485();
-            ~PowerRS485();
+            ~PowerRS485() = default;
 
         private:
-        RS485Provider *rs485;
+        RS485Driver *rs485;
 
         void publishMotorInfo(uint8_t cmd, std::vector<float> data);
