+++ src/main.cpp
@@ -1,3 +1,4 @@
+#include "rs485_port_manager/RS485Driver.hpp"
 #include "rs485_port_manager/RS485Provider.hpp"
 #include "rs485_port_manager/ActuatorRS485.hpp"
 #include "rs485_port_manager/KillMissionRS485.hpp"
@@ -11,7 +12,8 @@
 int main(int argc, char *argv[])
 {
     rclcpp::init(argc, argv);
-    auto rs485 = rs485_port_manager::RS485Provider::GetInstance();
+    auto rs485 = rs485_port_manager::RS485Driver::GetInstance();
+    auto providerRS485 = std::make_shared<rs485_port_manager::RS485Provider>();
     auto killRS485 = std::make_shared<rs485_port_manager::KillMissionRS485>();
     auto powerRS485 = std::make_shared<rs485_port_manager::PowerRS485>();
     auto ioRS485 = std::make_shared<rs485_port_manager::ActuatorRS485>();
@@ -28,6 +30,7 @@ int main(int argc, char *argv[])
     rs485->Start();
     rclcpp::executors::MultiThreadedExecutor executor;
     executor.add_node(killRS485);
+    executor.add_node(providerRS485);
     executor.add_node(powerRS485);
     executor.add_node(ioRS485);
     executor.spin();
